local-node: save local signal with neighbor-report
Save the signal of the current connected node when creating or saving
beacon-report data.

This way, the timeout for this reported data can be extended as long as
the signal strength of the current connection does not differ to much.

This feature is not yet implemented, but this change is necessary for
preperation.

Signed-off-by: David Bauer <mail@david-bauer.net>